Extract reusable code into utils
