+++ utils/index.ts
@@ -4,3 +4,5 @@ export * from './locations';
 export * from './select';
 export * from './strings';
 export * from './themes';
+export * from './numbers';
+export * from './values';
