+++ utils/numbers.ts
@@ -1,0 +1,12 @@
+export const formatNumber = (value: number): string => {
+  if (value >= 1000000000) {
+    return `${(value / 1000000000).toFixed(2)}b`;
+  }
+  if (value >= 1000000) {
+    return `${(value / 1000000).toFixed(2)}m`;
+  }
+  if (value >= 1000) {
+    return `${(value / 1000).toFixed(2)}k`;
+  }
+  return `${value.toFixed(2)}`;
+};
